# libstore: Use `boost::regex` for GC root discovery