@edolstra 2025-05-07
Please check that this doesn't cause a dependency on `libicu` (see https://github.com/NixOS/nix/pull/7762).