@Mic92 2025-05-13
> Please check that this doesn't cause a dependency on `libicu` (see #7762).

@xokdvium have you checked this?