@Mic92 2025-05-18
> @edolstra ping wrt `libicu` dependency on master. If you'd like I could split [f3090ef](https://github.com/NixOS/nix/commit/f3090ef7033c9bdc04beacfbb128c688cfa40fee) into a separate PR if that closure size reduction warrants a backport.

Unless there is someone else interested, I would just backport this to 2.29.